Allow FFmpeg if exists at /usr/local/bin/ffmpeg. 3.0.147

+++ trunk/auto/depends.sh
@@ -484,7 +484,7 @@ fi
 # live transcoding, ffmpeg-4.1, x264-core157, lame-3.99.5, libaacplus-2.0.2.
 #####################################################################################
 # Always link the ffmpeg tools if exists.
-if [[ -f /usr/local/bin/ffmpeg && ! -f ${SRS_OBJS}/ffmpeg/bin/ffmpeg ]]; then
+if [[ -f /usr/local/bin/ffmpeg && ! -f ${SRS_OBJS}/${SRS_PLATFORM}/ffmpeg ]]; then
     mkdir -p ${SRS_OBJS}/${SRS_PLATFORM}/ffmpeg/bin &&
     ln -sf /usr/local/bin/ffmpeg ${SRS_OBJS}/${SRS_PLATFORM}/ffmpeg/bin/ffmpeg &&
     (cd ${SRS_OBJS} && rm -rf ffmpeg && ln -sf ${SRS_PLATFORM}/ffmpeg)
@@ -493,11 +493,16 @@ if [ $SRS_FFMPEG_TOOL = YES ]; then
     if [[ -f ${SRS_OBJS}/${SRS_PLATFORM}/ffmpeg/bin/ffmpeg ]]; then
         echo "ffmpeg-4.1 is ok.";
     else
-        echo "no ffmpeg found, please use srs-docker or --without-ffmpeg";
+        echo "Warning: No FFmpeg found at /usr/local/bin/ffmpeg";
+        echo "    please copy it from srs-docker";
+        echo "    or download from http://ffmpeg.org/download.html";
+        echo "    or disable it by --without-ffmpeg";
         exit -1;
     fi
     # Always update the links.
-    (cd ${SRS_OBJS} && rm -rf ffmpeg && ln -sf ${SRS_PLATFORM}/ffmpeg)
+    if [[ -f ${SRS_OBJS}/${SRS_PLATFORM}/ffmpeg ]]; then
+        (cd ${SRS_OBJS} && rm -rf ffmpeg && ln -sf ${SRS_PLATFORM}/ffmpeg)
+    fi
 fi
 
 #####################################################################################
